requirements.txt
# AWS Bedrock
boto3>=1.34.0
botocore>=1.34.0

# Vector Database
chromadb>=0.4.22

# PDF Processing
pypdf>=3.17.0
Pillow>=10.1.0

# Data Processing
pandas>=2.1.4
numpy>=1.26.2

# Configuration
pyyaml>=6.0.1
python-dotenv>=1.0.0

# Utilities
jsonlines>=4.0.0
tqdm>=4.66.1

# Testing
pytest>=7.4.3
pytest-cov>=4.1.0

# Logging
loguru>=0.7.2

#Progress
tqdm

cryptography